+++ examples/multiwindow.rs
@@ -1,5 +1,3 @@
-extern crate glutin;
-
 mod support;
 
 use glutin::ContextTrait;
@@ -51,9 +49,9 @@ fn main() {
         for (index, window) in windows.values().enumerate() {
             let mut color = [1.0, 0.5, 0.7, 1.0];
             color.swap(0, index % 3);
-            let _ = unsafe { window.0.make_current() };
+            unsafe { window.0.make_current().unwrap() };
             window.1.draw_frame(color);
-            let _ = window.0.swap_buffers();
+            window.0.swap_buffers().unwrap();
         }
     }
 }