+++ src/os/macos.rs
@@ -1,17 +1,17 @@
 #![cfg(target_os = "macos")]
 
+use crate::os::ContextTraitExt;
+use crate::Context;
+
 pub use winit::os::macos::ActivationPolicy;
 pub use winit::os::macos::MonitorIdExt;
 pub use winit::os::macos::WindowBuilderExt;
 pub use winit::os::macos::WindowExt;
 
-use os::ContextTraitExt;
-use Context;
-
-use std::os::raw::c_void;
+use std::os::raw;
 
 impl ContextTraitExt for Context {
-    type Handle = *mut c_void;
+    type Handle = *mut raw::c_void;
 
     #[inline]
     unsafe fn raw_handle(&self) -> Self::Handle {
@@ -19,7 +19,7 @@ impl ContextTraitExt for Context {
     }
 
     #[inline]
-    unsafe fn get_egl_display(&self) -> Option<*const c_void> {
+    unsafe fn get_egl_display(&self) -> Option<*const raw::c_void> {
         None
     }
 }